Get generic docstring utils back from yaml_utils.py to utils.py
